[Diffusion] feat: seperate FlashAttention2 Backend, Hip use FA2 for default
