+++ python/sglang/multimodal_gen/runtime/layers/attention/backends/flash_attn.py
@@ -11,18 +11,8 @@ from sglang.srt.utils import is_hip
 
 _is_hip = is_hip()
 
-if _is_hip:
-    try:
-        from flash_attn import flash_attn_func  # Use the flash attention v2 function
-    except ImportError:
-
-        def _unsupported(*args, **kwargs):
-            raise ImportError(
-                "flash-attn is not installed. Please install it, e.g., `pip install flash-attn`."
-            )
-
-        flash_attn_func = _unsupported
-else:
+# hip sgl_kernel.flash_attn cannot import flash_attn_varlen_func
+if not _is_hip:
     try:
         from sgl_kernel.flash_attn import flash_attn_varlen_func
 
@@ -139,26 +129,17 @@ class FlashAttentionImpl(AttentionImpl):
             max_seqlen_q = query.shape[1]
             max_seqlen_k = key.shape[1]
 
-        if _is_hip:
-            output = flash_attn_func(
-                q=query,  # type: ignore[no-untyped-call]
-                k=key,
-                v=value,
-                softmax_scale=self.softmax_scale,
-                causal=self.causal,
-            )
-        else:
-            output = flash_attn_func(
-                q=query,  # type: ignore[no-untyped-call]
-                k=key,
-                v=value,
-                cu_seqlens_q=None,
-                cu_seqlens_k=None,
-                max_seqlen_q=max_seqlen_q,
-                max_seqlen_k=max_seqlen_k,
-                softmax_scale=self.softmax_scale,
-                causal=self.causal,
-                return_softmax_lse=return_softmax_lse,
-                ver=fa_ver,
-            )
+        output = flash_attn_func(
+            q=query,  # type: ignore[no-untyped-call]
+            k=key,
+            v=value,
+            cu_seqlens_q=None,
+            cu_seqlens_k=None,
+            max_seqlen_q=max_seqlen_q,
+            max_seqlen_k=max_seqlen_k,
+            softmax_scale=self.softmax_scale,
+            causal=self.causal,
+            return_softmax_lse=return_softmax_lse,
+            ver=fa_ver,
+        )
         return output
