+++ python/sglang/multimodal_gen/runtime/layers/attention/backends/flash_attn_2.py
@@ -10,9 +10,18 @@ from sglang.multimodal_gen.runtime.layers.attention.backends.attention_backend i
     AttentionMetadata,
     AttentionMetadataBuilder,
 )
-from sglang.multimodal_gen.runtime.layers.attention.backends.flash_attn import (
-    flash_attn_func,
-)
+
+try:
+    from flash_attn import flash_attn_func  # Use the flash attention v2 function
+except ImportError:
+
+    def _unsupported(*args, **kwargs):
+        raise ImportError(
+            "flash-attn is not installed. Please install it, e.g., `pip install flash-attn`."
+        )
+
+    flash_attn_func = _unsupported
+
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
 
 logger = init_logger(__name__)
@@ -39,7 +48,7 @@ class FlashAttention2Backend(AttentionBackend):
 
     @staticmethod
     def get_builder_cls() -> type["AttentionMetadataBuilder"]:
-        raise NotImplementedError
+        return None
 
 
 class FlashAttention2Impl(AttentionImpl):
@@ -68,10 +77,6 @@ class FlashAttention2Impl(AttentionImpl):
             q=query,  # type: ignore[no-untyped-call]
             k=key,
             v=value,
-            cu_seqlens_q=None,
-            cu_seqlens_k=None,
-            max_seqlen_q=None,
-            max_seqlen_k=None,
             softmax_scale=self.softmax_scale,
             causal=self.causal,
         )
